Collect extractors into `extract` module
