+++ rust-runtime/aws-smithy-http-server/src/extension.rs
@@ -3,35 +3,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-// This code was copied and then modified from Tokio's Axum.
-
-/* Copyright (c) 2021 Tower Contributors
- *
- * Permission is hereby granted, free of charge, to any
- * person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the
- * Software without restriction, including without
- * limitation the rights to use, copy, modify, merge,
- * publish, distribute, sublicense, and/or sell copies of
- * the Software, and to permit persons to whom the Software
- * is furnished to do so, subject to the following
- * conditions:
- *
- * The above copyright notice and this permission notice
- * shall be included in all copies or substantial portions
- * of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF
- * ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED
- * TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT
- * SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
- * CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR
- * IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
- */
-
 //! Extension types.
 //!
 //! Extension types are types that are stored in and extracted from _both_ requests and
@@ -50,14 +21,12 @@
 
 use std::ops::Deref;
 
-use http::StatusCode;
 use thiserror::Error;
 
-use crate::{
-    body::{empty, BoxBody},
-    request::{FromParts, RequestParts},
-    response::IntoResponse,
-};
+use crate::request::RequestParts;
+
+pub use crate::extract::extension::Extension;
+pub use crate::extract::extension::MissingExtension;
 
 /// Extension type used to store information about Smithy operations in HTTP responses.
 /// This extension type is set when it has been correctly determined that the request should be
@@ -148,49 +117,6 @@ impl Deref for RuntimeErrorExtension {
 
     fn deref(&self) -> &Self::Target {
         &self.0
-    }
-}
-
-/// Generic extension type stored in and extracted from [request extensions].
-///
-/// This is commonly used to share state across handlers.
-///
-/// If the extension is missing it will reject the request with a `500 Internal
-/// Server Error` response.
-///
-/// [request extensions]: https://docs.rs/http/latest/http/struct.Extensions.html
-#[derive(Debug, Clone)]
-pub struct Extension<T>(pub T);
-
-impl<T> Deref for Extension<T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-/// The extension has not been added to the [`Request`](http::Request) or has been previously removed.
-#[derive(Debug, Error)]
-#[error("the `Extension` is not present in the `http::Request`")]
-pub struct MissingExtension;
-
-impl<Protocol> IntoResponse<Protocol> for MissingExtension {
-    fn into_response(self) -> http::Response<BoxBody> {
-        let mut response = http::Response::new(empty());
-        *response.status_mut() = StatusCode::INTERNAL_SERVER_ERROR;
-        response
-    }
-}
-
-impl<Protocol, T> FromParts<Protocol> for Extension<T>
-where
-    T: Send + Sync + 'static,
-{
-    type Rejection = MissingExtension;
-
-    fn from_parts(parts: &mut http::request::Parts) -> Result<Self, Self::Rejection> {
-        parts.extensions.remove::<T>().map(Extension).ok_or(MissingExtension)
     }
 }
